+++ IsekaiMod/Content/Deities/Aqua.cs
@@ -109,8 +109,8 @@ namespace IsekaiMod.Content.Deities
             }));
 
             // Add to Deity Selection
-            var DeitySelection = Resources.GetBlueprint<BlueprintFeatureSelection>("59e7a76987fe3b547b9cce045f4db3e4");
-            DeitySelection.m_AllFeatures = DeitySelection.m_AllFeatures.AddToArray(GoddessAquaFeature.ToReference<BlueprintFeatureReference>());
+            var IsekaiDeitySelection = Resources.GetModBlueprint<BlueprintFeatureSelection>("IsekaiDeitySelection");
+            IsekaiDeitySelection.m_AllFeatures = IsekaiDeitySelection.m_AllFeatures.AddToArray(GoddessAquaFeature.ToReference<BlueprintFeatureReference>());
         }
     }
 }